+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 1. Created initial folder structure & files
 2. Added User and Thought models
 3. Added additional validation to email & reduced regex complexity
+4. Added Reaction model